youtube_summary_v1/requirements.txt
openai-whisper
django
pscyopg2
pytube
openai
sqlite 
